# openapi.yaml
openapi: 3.0.0
info:
  version: 0.0.1-beta
  title: storage-management-system
servers:
  - url: 'http://localhost:8080'
paths:
  /storage:
    get:
      summary: List all storage
      operationId: listStorage
      tags:
        - Storage
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 200
                message: List Storages Success
                data:
                - id: 2mgm35mm2135m1m5mh
                  releaseName: happy-tree
                  type: minio
                  config:
                    accessKey: aclab
                    secretKey: csie1226
                    persistence.size: 2G
                    service.type: NodePort
                    resources.limits.memory: 512Mi
                  endpoint: http://140.124.181.74:32296
                  status: updating
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 500
                message: Internal Server Error
    post:
      summary: Create a storage
      operationId: createStorage
      tags:
        - Storage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Storage'
            example:
              type: minio
              config:
                accessKey: aclab
                secretKey: csie1226
                persistence.size: 2G
                service.type: NodePort
                service.nodePort: ""
                resources.limits.memory: 512Mi
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 201
                message: Create Storage Success
                data:
                  id: 2mgm35mm2135m1m5mh
                  releaseName: happy-tree
                  type: minio
                  config:
                    accessKey: aclab
                    secretKey: csie1226
                    persistence.size: 2G
                    service.type: NodePort
                    service.nodePort: ""
                    resources.limits.memory: 512Mi
                  endpoint: http://140.124.181.74:32296
                  status: updating
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 500
                message: Internal Server Error
  '/storage/{releaseName}':
    get:
      summary: Info for a specific storage
      operationId: showStorageById
      tags:
        - Storage
      parameters:
        - name: releaseName
          in: path
          required: true
          description: The id of the storage to retrieve
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 200
                message: Get Storage Success
                data:
                  id: 2mgm35mm2135m1m5mh
                  releaseName: happy-tree
                  type: minio
                  config:
                    accessKey: aclab
                    secretKey: csie1226
                    persistence.size: 2G
                    service.type: NodePort
                    resources.limits.memory: 512Mi
                  endpoint: http://140.124.181.74:32296
                  status: updating
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 500
                message: Internal Server Error
    delete:
      summary: Delete a specific storage
      operationId: deleteStorageById
      tags:
        - Storage
      parameters:
        - name: releaseName
          in: path
          required: true
          description: The id of the storage to retrieve
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 200
                message: Delete Storage Success
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 500
                message: Internal Server Error
components:
  schemas:
    Storage:
      required:
        - type
        - config
      properties:
        id:
          type: string
        releaseName:
          type: string
        type:
          type: string
        config:
          type: object
        endpoint:
          type: string
        status:
          type: string
    Response:
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
        error:
          type: string